pyproject.toml
[tool.poetry]
name = "compass"
version = "0.1.0"
description = "Metrics and data to guide our engineering"
authors = ["Bukukas devops <devops@bukukas.io>"]

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.55"
prometheus-client = "^0.11.0"
gunicorn = "^20.1.0"
kopf = {extras = ["full-auth"], version = "^1.33.0"}
tenacity = "^8.0.1"
pykube-ng = {extras = ["gcp"], version = "^21.3.0"}
elasticsearch = {version = "<7.11", extras = ["async"]}
sentry-sdk = "^1.3.1"
elasticsearch-dsl = "^7.4.0"
arrow = "^1.1.1"

[tool.poetry.dev-dependencies]
prospector = {extras = ["with_everything"], version = "^1.3.1"}
jedi = {version = "^0.18.0", optional = true}
black = "^21.7b0"
ipython = {version = "^7.25.0", optional = true}
pudb = {version = "^2021.1", optional = true}
pytest-kind = "^21.1.3"
pytest = "^6.2.5"
types-PyYAML = "^5.4.10"
types-pytz = "^2021.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
